Change data state model
The state data has been changed from date based to only state.app_date
based.
All the app is based on the state.app_date. The initial date in cache is
set to setDefaultDate().
A change in the initial date of the state.app_date changes all the data
on the app.